recycler view is ready and the users are fetched

+++ ServerInfoViewer/app/src/main/java/com/example/serverinfoviewer/domain/entities/GeoData.kt
@@ -1,6 +1,6 @@
 package com.example.serverinfoviewer.domain.entities
 
 data class GeoData(
-    var lat: Float = 0f,
-    var lng: Float = 0f
+    var lat: Double = 0.0,
+    var lng: Double = 0.0
 )
